Fix: Terms 페이지 하나로 통합
